Rename workflow controller/reconciler
