+++ pkg/operator/reconciler/run/reconciler.go
@@ -1,4 +1,4 @@
-package workflow
+package run
 
 import (
 	"context"
